pyproject.toml
[project]
name = "nutrient-collab-test"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = ["flask", "gunicorn", "requests", "pyjwt", "cryptography"]
